tests/e2e/requirements_test.txt
pytest-playwright==0.3.0
requests==2.28.2
